1. 支持自动化创建新项目 2. 解决警告

+++ XBase/src/XPublic/XTools/XHTTP/XHttp.h
@@ -67,7 +67,7 @@ private:
 			pos++;
 		}
 		str = str.substr(pos, str.length() - pos);
-		pos = str.find(_sig);
+		pos = static_cast<int>(str.find(_sig));
 		if (pos == -1)
 		{
 			res = str.substr(0, str.length());
